libhimd: Move himd_writemp3() from himdcli to libhimd
